// Main/MG-Mod/types/models/spt/config/IQuestConfig.ts
export interface IQuestConfig extends IBaseConfig {
    kind: "spt-quest";
    // Hours to get/redeem items from quest mail keyed by profile type
    mailRedeemTimeHours: Record<string, number>;
    questTemplateIds: IPlayerTypeQuestIds;
    /** Show non-seasonal quests be shown to player */
    showNonSeasonalEventQuests: boolean;
    eventQuests: Record<string, IEventQuestData>;
    repeatableQuests: IRepeatableQuestConfig[];
    locationIdMap: Record<string, string>;
    bearOnlyQuests: string[];
    usecOnlyQuests: string[];
}

export interface IPlayerTypeQuestIds {
    pmc: IQuestTypeIds;
    scav: IQuestTypeIds;
}

export interface IQuestTypeIds {
    Elimination: string;
    Completion: string;
    Exploration: string;
}

export interface IEventQuestData {
    name: string;
    season: SeasonalEventType;
    startTimestamp: number;
    endTimestamp: number;
    yearly: boolean;
}

export interface IRepeatableQuestConfig {
    id: string;
    name: string;
    side: string;
    types: string[];
    resetTime: number;
    numQuests: number;
    minPlayerLevel: number;
    rewardScaling: IRewardScaling;
    locations: Record<ELocationName, string[]>;
    traderWhitelist: ITraderWhitelist[];
    questConfig: IRepeatableQuestTypesConfig;
    /** Item base types to block when generating rewards */
    rewardBaseTypeBlacklist: string[];
    /** Item tplIds to ignore when generating rewards */
    rewardBlacklist: string[];
    rewardAmmoStackMinSize: number;
    freeChangesAvailable: number;
    freeChanges: number;
}

export interface IRewardScaling {
    levels: number[];
    experience: number[];
    roubles: number[];
    gpCoins: number[];
    items: number[];
    reputation: number[];
    rewardSpread: number;
    skillRewardChance: number[];
    skillPointReward: number[];
}

export interface ITraderWhitelist {
    traderId: string;
    questTypes: string[];
    rewardBaseWhitelist: string[];
    rewardCanBeWeapon: boolean;
    weaponRewardChancePercent: number;
}

export interface IRepeatableQuestTypesConfig {
    Exploration: IExploration;
    Completion: ICompletion;
    Pickup: IPickup;
    Elimination: IEliminationConfig[];
}